Upgrade to v5.0.0
This version is a major rewrite of query object, and parsing process.

Rather than the previous Python dict format, which introduced limited
complex functionality, the new Query object allows for endless complex
forms of queries. It also solves the major problem with extracting the
query attrs from the object by introducing QueryAttrList which acts as a
list of all possible values of a query attr.

With this change, updated `permissions` sets now has ability to define
query as a Python list in order to make use of the `or` condition now
available with the new Query object.

This version also introduced `doc` opers which are set of opers that can
be used with `update` method in order to update fields within the `doc`
object without reading them, manipulating them and again push the
updated value to `update` method. Introduced opers are `$add`, `$push`,
`$pushUnique`, and `$pull`.

This version also rewrites `query_args` and `doc_args` and reformats
them to new format that has better developer-expereince and allows for
multiple optional attrs sets.